Make type verification more strict
The previous behavior was inconsistent as adding two scales of the same fundamental unit would fail while adding two scales of a compound unit would succeed
There's no good way to resolve this inconsistency in the opposite direction, as there's no record that (say) meters and kilometers are more related than meters and feet are.
This has the side effect of making the actual equality comparison in __eq__ irrelevant, as due to flyweight, `self is not other and type(other) is type(self)` implies `self.value != other.value`.